Add simplecov (locally)
Adding simplecov revealed that the lib/reformat.rb file, as a script
using optparse, doesn't readily fall into code coverage.  We were
still testing the functionality, by issuing command-line calls in the spec,
so we knew that it worked but couldn't see in reports that it was covered.

Adding simplecov influenced the change to move the non-optparse code
from the script into lib/file_interface.rb, which was a class and could then
be tested and code-coveraged exhaustively.

And we've still got the reformat_spec, renamed here cli_spec, which checks that
the bits in the optparse script file still work as expected, even if we
can't claim coverage for them.