Show errors in action summary page
Because rspec errors don't show as stderr, and there was a parsing
error in passing stdout as well, we're just passing stderr.

So the summary message will be "failed before it got to SimpleCov"
if an actual test failed, in which case click in to run-tests
to see which one failed, or it'll be the SimpleCov error showing
the test coverage is below the minimum threshold.